Move basic types to `mw/utils`
